+++ packages/xchain-crypto/src/crypto.ts
@@ -1,6 +1,6 @@
 import * as bip39 from 'bip39'
 import crypto from 'crypto'
-import { blake256 } from 'foundry-primitives'
+import { blake256 } from 'foundry-primitives-xchainjs'
 import { v4 as uuidv4 } from 'uuid'
 
 import { pbkdf2Async } from './utils'