+++ packages/xchain-bitcoin/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.22.0 (2023-05-02)
+
+## Update
+
+- update rollup config and axios to the latest
+
 # v.0.21.4 (2023-04-11)
 
 ## Add