+++ packages/xchain-mayachain/CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.2.0 (2023-05-02)
+
+## Update
+
+- update rollup config and axios to the latest
+- update `@types/big.js`
+
 # v0.1.2 (2023-04-04)
 
 ## add