+++ packages/xchain-utxo-providers/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.2.0 (2023-05-02)
+
+## Update
+
+- update rollup config and axios to the latest
+
 # v.0.1.2 (2023-4-11)
 
 ## Fix